+++ components/Announcement.jsx
@@ -1,43 +1,26 @@
 import React from "react";
-import Card from "react-bootstrap/Card";
+import { FaCircle } from "react-icons/fa";
 
-const Announcement = ({ color, title, date, time, location }) => {
+const Announcement = ({ background, text, title, date, time, location }) => {
   return (
-    <div>
-      <Card
-        className="w-fit h-fit !border-none !rounded-3xl font-lexend"
-        style={{
-          backgroundImage: `linear-gradient(to right,
-         ${color} 0% 6%,
-        black 6% 7.5%,
-         white 7.5% 100%)`,
-        }}
-      >
-        <Card.Body className="ml-8 flex-col justify-evenly">
-          <p className="text-4xl font-bold text-acm-black">{title}</p>
-          <div className="flex mb-2">
-            <div
-              className={`w-6 h-6 self-center flex-none rounded-full mr-2`}
-              style={{ backgroundColor: `${color}` }}
-            ></div>
-            <div className="text-acm-black text-xl">date: {date}</div>
-          </div>
-          <div className="flex mb-2">
-            <div
-              className={`w-6 h-6 self-center flex-none rounded-full mr-2`}
-              style={{ backgroundColor: `${color}` }}
-            ></div>
-            <div className="text-acm-black text-xl">time: {time}</div>
-          </div>
-          <div className="flex">
-            <div
-              className={`w-6 h-6 self-center flex-none rounded-full mr-2`}
-              style={{ backgroundColor: `${color}` }}
-            ></div>
-            <div className="text-acm-black text-xl">location: {location}</div>
-          </div>
-        </Card.Body>
-      </Card>
+    <div
+      className={`flex justify-end items-center my-3 w-11/12 ${background} border-4 border-black rounded-3xl`}
+    >
+      <div className="w-11/12 bg-acm-white border-l-4 border-acm-black rounded-r-3xl font-lexend p-2">
+        <p className="text-3xl font-semibold mx-2">{title}</p>
+        <div className="flex justify-start items-center text-xl">
+          <FaCircle className={`${text}  mx-2`} />
+          date: {date}
+        </div>
+        <div className="flex justify-start items-center text-xl">
+          <FaCircle className={`${text} mx-2`} />
+          time: {time}
+        </div>
+        <div className="flex justify-start items-center text-xl">
+          <FaCircle className={`${text}  mx-2`} />
+          location: {location}
+        </div>
+      </div>
     </div>
   );
 };
